code/requirements.txt
numpy==1.15.4
opencv_python==4.1.1.26
scikit_image==0.15.0
Pillow==7.0.0
scikit_learn==0.22.1
skimage==0.0
pytorch==1.13